Classes/CICD/automate-deploy/requirements.txt
boto3==1.35.48
botocore==1.35.48
exceptiongroup==1.2.2
iniconfig==2.0.0
jmespath==1.0.1
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.3
six==1.16.0
tomli==2.0.2
urllib3==1.26.20
